Add an attack sound, and redo how we make sounds.
Instead of issuing sounds immediately, we queue them in the world,
and they are played when the world is next drawn.  This lets us keep
client updating pure.

It may be nice to change this further, so that sounds are combined
with animation (e.g., start playing this when you get to frame X).